getDeviceLocalIPAsString: keep error context
Here's one take on why this is important:
https://www.yegor256.com/2015/12/01/rethrow-exceptions.html

The basic idea is to help whoever has to deal with these errors by
giving them more useful information. In this case, we capture the errors
(when available) and keep the original messages.

For example, this means we get to keep the stack-trace in the error, a
useful debugging aid (where did this error come from?).

For more, see the discussion at
https://github.com/intuit/user-data-for-fraud-prevention/issues/31